+++ changelog.md
@@ -3,6 +3,7 @@
 * Synced codebase, 1.16.5, 1.18.2 and 1.19.2 are now nearly identical again
 * Fixed offhand slot occluding steampunk armor
 * Steampunk armor has now 10% explosion resistance instead of 20%, reduced leggings armor by 1
+* Steampunk HUD Overlay also works on Forge now
 
 # 1.1.9